+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/psi/impl/xpath/XPathInlineFunctionExprPsiImpl.kt
@@ -76,7 +76,11 @@ class XPathInlineFunctionExprPsiImpl(node: ASTNode) :
             return when {
                 context.elementType === XPathTokenType.THIN_ARROW -> context
                 blockOpen?.isEmptyEnclosedExpr == true -> blockOpen
-                else -> variadicParameter ?: context
+                else -> when (val parameter = variadicParameter) {
+                    null -> context
+                    is XPathParam -> context
+                    else -> parameter
+                }
             }
         }